requirements.txt
torch==2.0.1
transformers==4.35.0
nltk==3.8.1
sentencepiece==0.1.99
PyPDF2==3.0.1
python-docx==1.1.0
python-pptx==0.6.21
pdfplumber==0.10.3
pytesseract==0.3.10
pillow==10.1.0
pandas==2.1.3
openpyxl==3.1.2